test(ui): split gap-closing & more-function-coverage into component files
Finding F2.1, UI batch (part 2). These padding files mixed several components;
split/consolidated into behavior-named homes (coverage preserved or improved):
- gap-closing.test.tsx -> pwa-update-toast.test.tsx + (TaskTimer block consolidated)
- more-function-coverage.test.tsx -> TaskTimer block consolidated into a single
  task-timer.test.tsx; its filter tests duplicated filters.test.ts (dropped); its
  one unique get-set branch (getSyncStatus deviceId fallback with no config) moved
  to config.test.ts as a real assertion (replacing a tautological one).
- task-card-coverage.test.tsx -> task-card-states.test.tsx (rename; covers
  selection, dependency indicators, subtask progress).

Verified: full suite green (1848 pass); every module holds or improves baseline;
total coverage unchanged (86.86% lines).
