--- tests/ui/more-function-coverage.test.tsx
@@ -1,270 +1,0 @@
-/**
- * Additional tests to boost function and branch coverage.
- * Targets: TaskTimer (more branches), filter-panel, sync config get-set.
- */
-import { describe, it, expect, vi, beforeEach } from "vitest";
-import { render, screen } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
-import { createMockTask } from "@/tests/fixtures";
-
-// ---------------------------------------------------------------------------
-// Common mocks
-// ---------------------------------------------------------------------------
-
-vi.mock("@/lib/logger", () => ({
-  createLogger: () => ({
-    info: vi.fn(),
-    warn: vi.fn(),
-    error: vi.fn(),
-    debug: vi.fn(),
-  }),
-}));
-
-const mockHasRunningTimer = vi.fn().mockReturnValue(false);
-const mockGetRunningEntry = vi.fn().mockReturnValue(undefined);
-const mockFormatTimeSpent = vi.fn((m: number) => `${m}m`);
-const mockIsTaskSnoozed = vi.fn().mockReturnValue(false);
-const mockGetRemainingSnoozeMinutes = vi.fn().mockReturnValue(0);
-
-vi.mock("@/lib/tasks", () => ({
-  hasRunningTimer: (...args: unknown[]) => mockHasRunningTimer(...args),
-  getRunningEntry: (...args: unknown[]) => mockGetRunningEntry(...args),
-  formatTimeSpent: (...args: unknown[]) => mockFormatTimeSpent(...args),
-  isTaskSnoozed: (...args: unknown[]) => mockIsTaskSnoozed(...args),
-  getRemainingSnoozeMinutes: (...args: unknown[]) => mockGetRemainingSnoozeMinutes(...args),
-}));
-
-// ---------------------------------------------------------------------------
-// 1. TaskTimer — additional branches
-// ---------------------------------------------------------------------------
-
-describe("TaskTimer — full/default variant", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-    mockHasRunningTimer.mockReturnValue(false);
-    mockGetRunningEntry.mockReturnValue(undefined);
-  });
-
-  it("renders full variant with start button and clock icon", async () => {
-    const { TaskTimer } = await import("@/components/task-timer");
-
-    render(
-      <TaskTimer
-        task={createMockTask({ timeSpent: 15 })}
-        onStartTimer={vi.fn().mockResolvedValue(undefined)}
-        onStopTimer={vi.fn().mockResolvedValue(undefined)}
-      />
-    );
-
-    expect(screen.getByLabelText(/start timer/i)).toBeInTheDocument();
-    expect(screen.getByText("15m")).toBeInTheDocument();
-  });
-
-  it("renders full variant with running timer", async () => {
-    mockHasRunningTimer.mockReturnValue(true);
-    const tenMinAgo = new Date(Date.now() - 10 * 60 * 1000).toISOString();
-    mockGetRunningEntry.mockReturnValue({
-      id: "entry-1",
-      startedAt: tenMinAgo,
-    });
-
-    const { TaskTimer } = await import("@/components/task-timer");
-
-    render(
-      <TaskTimer
-        task={createMockTask({
-          timeSpent: 30,
-          timeEntries: [{ id: "entry-1", startedAt: tenMinAgo }],
-        })}
-        onStartTimer={vi.fn().mockResolvedValue(undefined)}
-        onStopTimer={vi.fn().mockResolvedValue(undefined)}
-      />
-    );
-
-    expect(screen.getByLabelText(/stop timer/i)).toBeInTheDocument();
-  });
-
-  it("renders with estimated minutes and over-estimate", async () => {
-    const { TaskTimer } = await import("@/components/task-timer");
-
-    render(
-      <TaskTimer
-        task={createMockTask({ timeSpent: 120, estimatedMinutes: 60 })}
-        onStartTimer={vi.fn().mockResolvedValue(undefined)}
-        onStopTimer={vi.fn().mockResolvedValue(undefined)}
-      />
-    );
-
-    // Should show both tracked and estimated times
-    expect(screen.getByText("120m")).toBeInTheDocument();
-    expect(screen.getByText("60m")).toBeInTheDocument();
-  });
-
-  it("calls onStartTimer when play is clicked in full variant", async () => {
-    const onStart = vi.fn().mockResolvedValue(undefined);
-    const { TaskTimer } = await import("@/components/task-timer");
-    const user = userEvent.setup();
-
-    render(
-      <TaskTimer
-        task={createMockTask({ timeSpent: 0 })}
-        onStartTimer={onStart}
-        onStopTimer={vi.fn().mockResolvedValue(undefined)}
-      />
-    );
-
-    await user.click(screen.getByLabelText(/start timer/i));
-    expect(onStart).toHaveBeenCalledWith("test-task-1");
-  });
-
-  it("calls onStopTimer when pause is clicked", async () => {
-    mockHasRunningTimer.mockReturnValue(true);
-    const fiveMinAgo = new Date(Date.now() - 5 * 60 * 1000).toISOString();
-    mockGetRunningEntry.mockReturnValue({
-      id: "entry-1",
-      startedAt: fiveMinAgo,
-    });
-
-    const onStop = vi.fn().mockResolvedValue(undefined);
-    const { TaskTimer } = await import("@/components/task-timer");
-    const user = userEvent.setup();
-
-    render(
-      <TaskTimer
-        task={createMockTask({
-          timeSpent: 10,
-          timeEntries: [{ id: "entry-1", startedAt: fiveMinAgo }],
-        })}
-        onStartTimer={vi.fn().mockResolvedValue(undefined)}
-        onStopTimer={onStop}
-      />
-    );
-
-    await user.click(screen.getByLabelText(/stop timer/i));
-    expect(onStop).toHaveBeenCalledWith("test-task-1");
-  });
-});
-
-// ---------------------------------------------------------------------------
-// 2. Filter-related tests — more branch coverage via filters module
-// ---------------------------------------------------------------------------
-
-describe("filter utilities", () => {
-  it("BUILT_IN_SMART_VIEWS has expected views", async () => {
-    const { BUILT_IN_SMART_VIEWS } = await import("@/lib/filters");
-    expect(BUILT_IN_SMART_VIEWS.length).toBeGreaterThan(0);
-    // Each built-in view should have a name and filter
-    for (const view of BUILT_IN_SMART_VIEWS) {
-      expect(view.name).toBeDefined();
-      expect(view.isBuiltIn).toBe(true);
-    }
-  });
-
-  it("applyFilters filters tasks correctly", async () => {
-    const { applyFilters } = await import("@/lib/filters");
-
-    const tasks = [
-      createMockTask({
-        id: "t1",
-        title: "Active Task",
-        completed: false,
-        tags: ["work"],
-      }),
-      createMockTask({
-        id: "t2",
-        title: "Done Task",
-        completed: true,
-        tags: ["personal"],
-      }),
-    ];
-
-    // Filter for active only
-    const active = applyFilters(tasks, { status: "active" });
-    expect(active.length).toBe(1);
-    expect(active[0].id).toBe("t1");
-
-    // Filter for completed
-    const completed = applyFilters(tasks, { status: "completed" });
-    expect(completed.length).toBe(1);
-    expect(completed[0].id).toBe("t2");
-  });
-
-  it("applyFilters with tag filter", async () => {
-    const { applyFilters } = await import("@/lib/filters");
-
-    const tasks = [
-      createMockTask({ id: "t1", tags: ["work", "frontend"] }),
-      createMockTask({ id: "t2", tags: ["personal"] }),
-    ];
-
-    const filtered = applyFilters(tasks, { tags: ["work"] });
-    expect(filtered.length).toBe(1);
-    expect(filtered[0].id).toBe("t1");
-  });
-
-  it("applyFilters with quadrant filter", async () => {
-    const { applyFilters } = await import("@/lib/filters");
-
-    const tasks = [
-      createMockTask({
-        id: "t1",
-        quadrant: "urgent-important",
-        urgent: true,
-        important: true,
-      }),
-      createMockTask({
-        id: "t2",
-        quadrant: "not-urgent-important",
-        urgent: false,
-        important: true,
-      }),
-    ];
-
-    const filtered = applyFilters(tasks, {
-      quadrants: ["urgent-important"],
-    });
-    expect(filtered.length).toBe(1);
-    expect(filtered[0].id).toBe("t1");
-  });
-
-  it("applyFilters with no filter returns all", async () => {
-    const { applyFilters } = await import("@/lib/filters");
-    const tasks = [createMockTask({ id: "t1" }), createMockTask({ id: "t2" })];
-
-    const filtered = applyFilters(tasks, {});
-    expect(filtered.length).toBe(2);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// 3. Sync config/get-set — more branches
-// ---------------------------------------------------------------------------
-
-describe("sync config get-set", () => {
-  beforeEach(async () => {
-    const { getDb } = await import("@/lib/db");
-    const db = getDb();
-    await db.syncMetadata.clear();
-  });
-
-  it("getSyncConfig returns null when no config exists", async () => {
-    const { getSyncConfig } = await import("@/lib/sync/config");
-    const config = await getSyncConfig();
-    // Returns null or the config — depends on initial migration state
-    // Just verify it doesn't throw
-    expect(config === null || typeof config === "object").toBe(true);
-  });
-
-  it("isSyncEnabled returns false when not configured", async () => {
-    const { isSyncEnabled } = await import("@/lib/sync/config");
-    const enabled = await isSyncEnabled();
-    expect(enabled).toBe(false);
-  });
-
-  it("getSyncStatus returns status object", async () => {
-    const { getSyncStatus } = await import("@/lib/sync/config");
-    const status = await getSyncStatus();
-    expect(typeof status.enabled).toBe("boolean");
-    expect(typeof status.pendingCount).toBe("number");
-  });
-});
